refactor(actionbar): move cancel and queue buttons into draggable container
